requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.20.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
httpx>=0.24.0
tavily-python>=0.3.0 